js/xpconnect/tests/chrome/test_ccdump.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=709162
-->
<window title="Mozilla Bug 709162"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=709162"
     target="_blank">Mozilla Bug 709162</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[
  /** Test for Bug 709162 **/

  var emptyListener = {
    QueryInterface: function QueryInterface(aIID) {
      if (aIID.equals(Components.interfaces.nsICycleCollectorListener) ||
          aIID.equals(Components.interfaces.nsISupports))
        return this;
      throw Components.results.NS_NOINTERFACE;
    },

    /* nsICycleCollectorListener */
    begin: function () {},
    noteRefCountedObject: function (addr, rc, descr) {},
    noteGCedObject: function (addr, marked, descr) {},
    noteEdge: function (addr, descr) {},
    beginResults: function () {},
    describeRoot: function (addr, known) {},
    describeGarbage: function (addr) {},
    end: function () {},
  };

  window.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
        .getInterface(Components.interfaces.nsIDOMWindowUtils)
        .cycleCollect(emptyListener);

  ok(true, "Dump cycle collector graph without crashing.");

  ]]>
  </script>
</window>
